Made Stargate spawn in proper position. Made modules take damage and SelfDestruct. Fixed various typecast warnings

+++ Entities/Ships/Player.cpp
@@ -135,7 +135,7 @@ void CPlayer::CreateMissile()
 	missileTimer = 0;
 	//TODO: Send CreateMissile message
 	int damage = 75;
-	damage *= (1.5f * missileLevel);
+	damage *= int(1.5f * missileLevel);
 	CreateProjectileMessage* msg = new CreateProjectileMessage(EntityType::Missile, position, size, rotation, damage, missileLevel );
 	msg->QueueMessage();
 }
